docs: Open-AD-Kit: Setup ADE for aarch64 and x86_64 hosts
The documentation doesn't describe clearly for how to setup ADE on
x86_64 host, and we need to consider to setup for both aarch64 and
x86_64 hosts.

This patch gives the commands for downloading ADE binary for aarch64 and
x86_64 versions, and then we can use the common commands to update ADE
binary and copy the binary to system path.

Signed-off-by: Leo Yan <leo.yan@linaro.org>